requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
mysql-connector-python
python-dotenv
xhtml2pdf
num2words
pandas
openpyxl
fastapi
uvicorn
python-dotenv
pydantic
mysql-connector-python
requests
openai
groq
python-dateutil
flask-cors
